Add Game Log
Potential Issues: Reserving a card is exposed strangely. We probably want to change it so that a random one
is a different action than a faceup one so we can not display an opponents randomly drafted card.
This is a extremely poor log we should pretty up the information we display