Rewrite Logging structure
This changes how logs are stored. Now each entry is a length 2 array, the first element is an
event type and the second is a payload. The first entry in the payload should be the userID of
who took the action though this is not a strict requirement. Prettifying is still a TODO